+++ apps/test-suite/tests/GLView.js
@@ -133,6 +133,53 @@ export async function test(
         gl.endFrameEXP();
       });
 
+      it(`draws a compressed texture`, async () => {
+        const gl = await getContextAsync();
+
+        const vert = gl.createShader(gl.VERTEX_SHADER);
+        gl.shaderSource(vert, vertexShader);
+        gl.compileShader(vert);
+        const frag = gl.createShader(gl.FRAGMENT_SHADER);
+        gl.shaderSource(frag, fragShader);
+        gl.compileShader(frag);
+
+        const program = gl.createProgram();
+        gl.attachShader(program, vert);
+        gl.attachShader(program, frag);
+        gl.linkProgram(program);
+        gl.useProgram(program);
+
+        const buffer = gl.createBuffer();
+        gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+        const verts = new Float32Array([-2, 0, 0, -2, 2, 2]);
+        gl.bufferData(gl.ARRAY_BUFFER, verts, gl.STATIC_DRAW);
+        const positionAttrib = gl.getAttribLocation(program, 'position');
+        gl.enableVertexAttribArray(positionAttrib);
+        gl.vertexAttribPointer(positionAttrib, 2, gl.FLOAT, false, 0, 0);
+
+        const texture = gl.createTexture();
+        gl.activeTexture(gl.TEXTURE0);
+        gl.bindTexture(gl.TEXTURE_2D, texture);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
+
+        gl.compressedTexImage2D(
+          gl.TEXTURE_2D,
+          0,
+          gl.COMPRESSED_R11_EAC,
+          32,
+          32,
+          0,
+          new Uint8Array([255, 0, 0, 255, 0, 255, 0, 255, 0, 0, 255, 255, 128, 128, 0, 255])
+        );
+        gl.uniform1i(gl.getUniformLocation(program, 'texture'), 0);
+
+        gl.clearColor(0, 0, 1, 1);
+        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
+        gl.drawArrays(gl.TRIANGLES, 0, verts.length / 2);
+        gl.endFrameEXP();
+      });
+
       it(`draws a texture with a TypedArray`, async () => {
         const gl = await getContextAsync();
 
@@ -174,6 +221,57 @@ export async function test(
           2,
           gl.RGBA,
           gl.UNSIGNED_BYTE,
+          new Uint8Array([255, 0, 0, 255, 0, 255, 0, 255, 0, 0, 255, 255, 128, 128, 0, 255])
+        );
+
+        gl.uniform1i(gl.getUniformLocation(program, 'texture'), 0);
+
+        gl.clearColor(0, 0, 1, 1);
+        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
+        gl.drawArrays(gl.TRIANGLES, 0, verts.length / 2);
+        gl.endFrameEXP();
+      });
+
+      it(`draws a compressed texture with a TypedArray`, async () => {
+        const gl = await getContextAsync();
+
+        const vert = gl.createShader(gl.VERTEX_SHADER);
+        gl.shaderSource(vert, vertexShader);
+
+        gl.compileShader(vert);
+        const frag = gl.createShader(gl.FRAGMENT_SHADER);
+        gl.shaderSource(frag, fragShader);
+        gl.compileShader(frag);
+
+        const program = gl.createProgram();
+        gl.attachShader(program, vert);
+        gl.attachShader(program, frag);
+        gl.linkProgram(program);
+        gl.useProgram(program);
+
+        const buffer = gl.createBuffer();
+        gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+        const verts = new Float32Array([-2, 0, 0, -2, 2, 2]);
+        gl.bufferData(gl.ARRAY_BUFFER, verts, gl.STATIC_DRAW);
+        const positionAttrib = gl.getAttribLocation(program, 'position');
+        gl.enableVertexAttribArray(positionAttrib);
+        gl.vertexAttribPointer(positionAttrib, 2, gl.FLOAT, false, 0, 0);
+
+        const texture = gl.createTexture();
+        gl.activeTexture(gl.TEXTURE0);
+        gl.bindTexture(gl.TEXTURE_2D, texture);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
+
+        // Use below to test using a `TypedArray` parameter
+        gl.compressedTexSubImage2D(
+          gl.TEXTURE_2D,
+          0,
+          32,
+          32,
+          2,
+          2,
+          gl.COMPRESSED_R11_EAC,
           new Uint8Array([255, 0, 0, 255, 0, 255, 0, 255, 0, 0, 255, 255, 128, 128, 0, 255])
         );